update jwt services structure to be its own file
